+++ src/harlequin/cli.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import sys
 from typing import Any, Callable
 
-import click
+import rich_click as click
 
 from harlequin import Harlequin
 from harlequin.adapter import HarlequinAdapter
@@ -12,6 +12,51 @@ if sys.version_info < (3, 10):
     from importlib_metadata import entry_points
 else:
     from importlib.metadata import entry_points
+
+# configure the rich click interface (mostly --help options)
+DOCS_URL = "https://harlequin.sh/docs/getting-started"
+GREEN = "#45FFCA"
+YELLOW = "#FEFFAC"
+PINK = "#FFB6D9"
+PURPLE = "#D67BFF"
+
+# general
+click.rich_click.USE_RICH_MARKUP = True
+click.rich_click.COLOR_SYSTEM = "truecolor"
+
+click.rich_click.STYLE_OPTIONS_TABLE_LEADING = 1
+click.rich_click.STYLE_OPTIONS_TABLE_BOX = "SIMPLE"
+click.rich_click.STYLE_OPTIONS_PANEL_BORDER = YELLOW
+click.rich_click.STYLE_USAGE = f"bold {YELLOW}"
+click.rich_click.STYLE_USAGE_COMMAND = "regular"
+click.rich_click.STYLE_HELPTEXT = "regular"
+click.rich_click.STYLE_OPTION = PINK
+click.rich_click.STYLE_ARGUMENT = PINK
+click.rich_click.STYLE_COMMAND = PINK
+click.rich_click.STYLE_SWITCH = GREEN
+
+# metavars
+click.rich_click.SHOW_METAVARS_COLUMN = False
+click.rich_click.APPEND_METAVARS_HELP = True
+click.rich_click.STYLE_METAVAR_APPEND = PURPLE
+click.rich_click.STYLE_METAVAR_SEPARATOR = PURPLE
+
+# errors
+click.rich_click.STYLE_ERRORS_SUGGESTION = "italic"
+click.rich_click.ERRORS_SUGGESTION = "Try 'harlequin --help' to view available options."
+click.rich_click.ERRORS_EPILOGUE = (
+    f"To learn more, visit [link={DOCS_URL}]{DOCS_URL}[/link]"
+)
+
+# define main option group (adapter options added to own groups below)
+click.rich_click.OPTION_GROUPS = {
+    "harlequin": [
+        {
+            "name": "Harlequin Options",
+            "options": ["--adapter", "--theme", "--limit", "--version", "--help"],
+        },
+    ]
+}
 
 
 def build_cli() -> click.Command:
@@ -83,13 +128,17 @@ def build_cli() -> click.Command:
         """
         This command starts the Harlequin IDE.
 
-        conn_str TEXT: One or more connection strings (or paths to local db files)
-        for databases to open with Harlequin.
+        [bold #FFB6D9]CONN_STR[/] [#D67BFF](TEXT MULTIPLE)[/][dim]: One or more
+        connection strings (or paths to local db files) for databases to open with
+        Harlequin.[/]
         """
         # prune the kwargs to only those that don't have their default arguments
         params = list(kwargs.keys())
         for k in params:
-            if ctx.get_parameter_source(k) == click.core.ParameterSource.DEFAULT:
+            if (
+                ctx.get_parameter_source(k)
+                == click.core.ParameterSource.DEFAULT  # type: ignore
+            ):
                 kwargs.pop(k)
 
         # load and instantiate the adapter
@@ -108,14 +157,19 @@ def build_cli() -> click.Command:
     # we load the options into a dict keyed by their name to de-dupe options
     # that may be passed by multiple adapters.
     options: dict[str, Callable[[click.Command], click.Command]] = {}
-    for adapter_cls in adapters.values():
+    for adapter_name, adapter_cls in adapters.items():
+        option_name_list: list[str] = []
         if adapter_cls.ADAPTER_OPTIONS is not None:
             for option in adapter_cls.ADAPTER_OPTIONS:
                 options.update({option.name: option.to_click()})
+                option_name_list.append(f"--{option.name}")
+        click.rich_click.OPTION_GROUPS["harlequin"].append(
+            {"name": f"{adapter_name} Adapter Options", "options": option_name_list}
+        )
 
     fn = inner_cli
     for click_opt in options.values():
-        fn = click_opt(fn)
+        fn = click_opt(fn)  # type: ignore
 
     return fn